Version I - BROKEN SELECT MODE
More styling for mode indicators. Added utility function for toggling classes. Added functions for Inner Mode. Attempted to add Select Mode but decided to try a simpler way to achieve the same goal (called List Mode, starting in Version I in the Main Branch of this repo). This code is buggy.
Keeping this branch here for completeness.